+++ src/StackTemplates/Cmd.hs
@@ -1,18 +1,19 @@
-{-# LANGUAGE OverloadedLabels #-}
-
 module StackTemplates.Cmd
     ( module X
     , Cmd (..)
     , toCmd
+    , Env
     ) where
 
 import           RIO
 
 import           StackTemplates.Cmd.Options as X
 import           StackTemplates.Cmd.Run     as X
+import           StackTemplates.Env         (Env)
 
 data Cmd
   = PrintVersion
+  | PrintHelp
   | FetchTplList Options
   | FetchRawTpl Text Options
   deriving (Show, Eq)
@@ -20,6 +21,7 @@ data Cmd
 toCmd :: Options -> Maybe Cmd
 toCmd opts
   | opts ^. #version = Just PrintVersion
+  | opts ^. #help    = Just PrintHelp
   | opts ^. #list    = Just $ FetchTplList opts
   | otherwise        = flip FetchRawTpl opts <$> path
   where