+++ src/StackTemplates/Env.hs
@@ -1,24 +1,22 @@
-{-# LANGUAGE DataKinds        #-}
-{-# LANGUAGE OverloadedLabels #-}
-{-# LANGUAGE TypeOperators    #-}
-{-# OPTIONS_GHC -fno-warn-orphans #-}
-
 module StackTemplates.Env where
 
 import           RIO
 import           RIO.Directory
+import           RIO.FilePath
 
 import           Data.Extensible
+import           Mix.Plugin.Logger ()
+
 
 type Env = Record
-  '[ "logger"   >: LogFunc
-   , "gh_token" >: ByteString
+  '[ "logger"      >: LogFunc
+   , "gh_token"    >: ByteString
+   , "with_update" >: Bool
+   , "only_link"   >: Bool
    ]
-
-instance HasLogFunc Env where
-  logFuncL = lens (view #logger) (\x y -> x & #logger `set` y)
 
 cacheTplsListFile :: MonadIO m => m FilePath
 cacheTplsListFile = do
   cacheDir <- getXdgDirectory XdgCache "stack-tpls"
-  pure $ cacheDir ++ "stack-teplates"
+  createDirectoryIfMissing True cacheDir
+  pure $ cacheDir </> "stack-teplates"